change parameters type

+++ src/modules/uORB/topics/aslctrl_parameters.h
@@ -182,13 +182,13 @@ struct aslctrl_parameters_s {
 	float CLSYSID_step; 				// excitation unit step size eg: 5, -5 etc. (in degrees), NOTE: sign implies the direction of the first step in a given maneuver
 	float CLSYSID_nom_pitch;			// nominal pitch-reference value
 	float CLSYSID_nom_roll;				// nominal roll-reference value
-	uint8_t CLSYSID_f_start;			// start frequency for chirp
-	uint8_t CLSYSID_f_end;				// end frequency for chirp
+	float CLSYSID_f_start;			// start frequency for chirp
+	float CLSYSID_f_end;				// end frequency for chirp
 	float CLSYSID_Fs; 					// sampling rate for chirp
 	uint8_t CLSYSID_maneuver; 			// maneuver selection: 2-1-1 (0), chirp (1)
 	uint8_t CLSYSID_ctrlinput; 			// control selection: pitch (0), roll (1)
 	uint8_t CLSYSID_repeat; 			// number of times to repeat sys id maneuver
-	uint8_t CLSYSID_settime;			// settling time before maneuver starts
+	float CLSYSID_settime;			// settling time before maneuver starts
 };
 
 /**
